@@ -1,7 +1,6 @@
-import ReactDOM from 'react-dom';
-import App from './App';
+import React from 'react'
+import ReactDOM from 'react-dom/client'
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
-)
+import App from './App'
+
+ReactDOM.createRoot(document.getElementById('root')).render(<App />)
